config: ignore incompatible errors with hugo 0.73.0
ERROR: You have configured output formats for 'taxonomy' in your site
configuration. In Hugo 0.73.0 we fixed these to be what most people
expect (taxonomy and term).  But this also means that your site
configuration may not do what you expect. If it is correct, you can
suppress this message by following the instructions below.

If you feel that this should not be logged as an ERROR, you can ignore
it by adding this to your site config:

    ignoreErrors = ["error-output-taxonomy"]

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>